api/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
fastapi==0.108.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
Mako==1.3.0
MarkupSafe==2.1.3
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
python-environ==0.4.54
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.32.0.post1
typing_extensions==4.9.0
uvicorn==0.25.0
